Refactor: 코드 리뷰 피드백 반영 (윤환) (#89)
* [#83] Refactor: Login 관련 UI 컴포넌트화

* [#83] Refactor: Login 폼 기능 분리

* [#83] Refactor: 로그인 버튼 onClick에서 onSubmit 방식으로 변경

* [#83] Refactor: event 매개변수 누락 수정

* [#83] Refactor: Login 페이지 개별 Footer 파일 제거
